requirements.txt
numpy==2.2.5
Pillow==11.2.1
torch==2.6.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
